requirements.txt
absl-py==2.1.0
cachetools==5.4.0
charset-normalizer==3.3.2
cycler==0.11.0
google-auth==2.33.0
google-auth-oauthlib==0.4.6
grpcio==1.62.3
h5py==2.10.0
idna==3.7
importlib-metadata==6.7.0
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.4.3
numpy==1.21.2
oauthlib==3.2.2
opencv-python==4.5.3.56
Pillow==8.2.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.7.1
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.62.2
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
